Add files via upload
This is a Jupyter Notebook for a project I did to analyze heart disease patients data, and manipulate and extract useful information to create a machine learning algorithm which can predeict whether or not a future patient may be at risk to heart disease or are more likely to be diagnosed with a heart disease. This is a project that makes use of Unsupervised Learning methodolgy and has taught me all the fundamentals on how to handle and manipulate data in a data science environment.