Add files via upload
In this project I used a regression model of machine learning to predict future prices of bulldozer equipment. This is a Supervised Learning Project in which I used various Python Data Science libraries, such as pandas, numpy, matplotlib and more to deduce and calculate my predictions for future bulldozer prices.